Implemented same laser cannon shooting as before
