+++ Assets/Prefabs/Weapons/CruiserCannons.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a2dfd1d5123214ae6985944ceb5e7be3
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
