+++ Assets/Scripts/Mechanics/Ships/Combat/Weapons/DirectionalWeapon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c52121adf67974e90b68d6638b8a26d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
