+++ Assets/Scripts/Mechanics/Ships/Combat/Weapons/LaserCannonArray.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b55412b650234452b9a58514f7cc9dd2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
